Initial commit
Server listens for a call to /micropub and assumes clients are
handling authz/authn with indyauth. Does proper validation of incoming
POST requests and returns a 200 OK if the validation request succeeds.
Still TODO:

- Use verification response to validate the action being taken is
authorized (e.g. inspect the 'me' and 'nonce' fields)
- Actually store both metadata from the verification of the client as
well as the actual payload (in case of a create). Maybe flat files to
start.
- Return the expected response body
- Make auth verification server configurable to match the auth service
the host website uses
request
- (later) put the content somewhere more permanent/easily queryable
- (later) handle update requests
- (later) handle media uploads
- (later) handle distinquishing various post types