Check that the token used is for correct 'me'
For now the check for the 'me' field is against a hardcoded URL. This is
a candidate for being extracted into configuration just like the other
consts in this server.

Also adds a rejection handler. This is used to ensure an incorrect 'me'
after the token validation results in a 401 instead of a 500, that
our other custom rejections are still 500s, and that any other rejection
type is passed along the Filter stack.