Add helper for generating URL slugs
There will be cases where a slug is not provided as part of creating a
new entry/post. When that is the case, we'll still want a slug to be
created - this is a first pass on doing so.

It first looks for a title/name of the post and uses that. If that does
not exist then it will use the first bit of the content. It will URL
encode the whole thing before returning.

This is meant to be used only if a slug isn't provided and so doesn't
take into account a slug property being sent as part of the micropub
creation POST. That isn't currently a supported property for this server
but will go on the TODO list.