Write posts into sqlite database
This adds a sqlite connection manager/pool as a dependency to our
micropub handler and does the appropriate set up for that connection in
main. We also add to the (now misnamed) verify auth method of the
micropub handler to deal with actual massaging of post data into the
database.

It's just shoved into the handler directly for now which is pretty ugly.
This whole server needs a good amount of refactoring and splitting of
responsibilities (and probably more modules) but it's working for now.
We do the post creation within a transaction - this is because we need
to first create the post using relevant data. We have a separate table
for categories/tags though and so then separately need to see which post
we just created by getting the biggest post ID (sqlite doesn't support
returning this after executing the query) and then create our entries in
the categories table that can reference the right post_id.