Implement post fetch handler v1
Begins extracting custom rejections into an errors module. Implements a
handler to deal with fetching posts given a slug. Returns a JSON
serialized version of the post if the slug is valid otherwise returns a
404.

Lot of pieces missing here still but this is a good first pass. Notably:

- Post categories are missing from the response, we need to query for
those separately
- JSON isn't really what we want to return in the end but it's good for
validating that this all works
- I'll likely want to implement some sort of view layer that takes a
full Post (including categories) and return a rendered/materialized view
suitable for humans

Separately from this project, I need to begin considering how to
integrate routing requests to this server from my website. Currently the
website is a static site with files pre-rendered on disk via the Pelican
static site generator. Maybe I can find a way to import the existing
posts?