Handle single category field or none at all
In testing various micropub clients to publish via this server, I
noticed a couple peculiarities around how different services utilized
the category field to publish tags.

I used micropublish.net and quill.p3k.io in my tests. I found that one
or both of them:

- Would send no category field at all if no categories were provided on
  the new entry. (Reasonably!)
- Would send category as a singular field ('category') instead of an
  array ('category[]') if only one category was provided.

Unfortunately, neither the querystring deserialization library I was using
(serde_qs) to integrate into Serde to get a struct from the query
string, nor its primary competitor, serde_urlencoded, supported
either of these use cases.

To support not having a category at all I would have needed the field's
type to be Option<Vec<String>> which I didn't really want. For this case
I was comfortable with an empty Vec signifying the lack of categories
that need to be saved.

I'm not sure if there was a workaround at all for the clients
alternating between 'category' and 'category[]'. I couldn't think of one
at least.

Instead of using those integrations and the nice automatic
deserialization into a struct, I dropped down a level and am using the
url crate to parse the query string myself into a new builder that
handles all of the above cases for categories. I added tests for these
two new cases as well.

The main downside is that it's slightly more code to maintain but at
least it's something I directly control and as a result I have fewer
dependencies (anoyhow not included... it's not strictly required) The
URL crate was already used under the hood by the serde_qs crate so it's
not a _new_ dep, just a new direct dep. Another downside is that I'll
have to do more custom parsing for some another use cases that I'll need
to support: nested objects ('content[html]'). Instead of making
'content' an object with an HTML field that serde would handle properly,
I'll need to handle it myself. That's OK though as it's another case
where content could be either 'content' or 'content[html]' (or the body
could even be JSON instead of form encoded, but that's another day)!

Tradeoffs, tradeoffs...