First-pass JSON content type micropub handler
Uses a Map<String, Vec<String>> to house properties dynamically which
was one of the issues I was running into before. This was all written
before today but had not been committed so I am not quite sure what, if
any, issues there are with it. It compiles but I don't remember if I
tested it or anything – committing so I have a starting point to
exercise the new content type handler.