Abstract away templating and factor out common base context settings
This removes direct references to the Tera template system types in the
handlers and insteads references a new internal Templates type. This
allows us to wrap up the template functionality that is needed into an
internal interface if needed and contains where we're exposed to Tera.

While at it, it also removes some duplication with the base context
settings that all of the pages need due to the base template file.