Initial Atom Feed support
Includes an atom template in source code here as it turns out that's not
part of the pelican template base. Ensures the relevant template
variables related to feeds are available.

For now, only provides a feed for all items (all.atom.xml) but could
provide more specialized feeds later on. It creates a feed out of all
items (no limit on feed entries) as I don't think it'll be a problem to
load all posts for a bit. Could change that later on. Additionally, says
the feed's last updated time is the most recent updated time of all the
posts loaded. If I ever put a limit on posts in the feed, I'll need
consider whether this method needs a revamp.

Unrelated, but provides an explicit 404 page body because Firefox (or
more likely some extension) was barfing on having no content to render.