Add support for browsing tag archives
Adds a new query to fetch all posts by a given tag name. Then uses that
query in a refactored archive handler that allows an optional tag. If
the tag is passed in, posts that match the tag (category) will be
returned, otherwise all posts are returned.

Additionally adds a template context variable so that the template's
header may be modified based on context.