refactor: break up MicropubHandler::verify_auth
verify_auth was doing more than verifying auth headers. We now have a
handle_post (as in HTTP POST method) that calls a simplified verify_auth
method that either verifies the authentication or returns a rejection.
We additionally have a new create_post method that will actually handle
parsing the body and inserting appropriate rows into the database. The
handle_post method then takes the slug that was created via create_post
and returns our HTTP response with the right Location header.

This is better organization for the MicropubHandler but also allows
for our (coming) import entry binary to call only the create_post part
of this handler to allow for locally importing posts from body content
without needing a full HTTP request.