feat: support mp-slug property to override default slug
This allows input payloads to specify the slug that they are published
to. This will be helpful for backfilling the posts that use a slug
format of just a post name with no date info in them. This will allow me
to maintain the posts at the same URLs as before.

Depending on how various micropub clients work, I may want to later
reconsider whether to append mp-slug to a date-based path.