fix: default datetime offset of -8 rather than +7
This was causing the show time to be ahead of what was stored. Times
were stored in UTC (or local time with some implicit offset for many
imported posts...), however, we would then parse with a fixed offset.
That hardcoded offset was for a timezone that ~none of the posts were
ever written in.

This is a stopgap until I come up with something more sophisticated for
storing and handling timezone offsets either on a date range or per-post
basis. Timezones are a pain ;).