refactor: remove itertools dep
We already have the posts sorted via our database query and the template
for our atom feed already depends on that fact. Not a stretch to
continue to assume that here as well. This allows us to remove the
itertools crate as we were only using the `sorted()` helper here and not
using the extensions anywhere else.

We could consider adding an assertion for the reverse sort property if
we were concerned this might change out from under us.