refactor: simplify db connection pool interactions
This commit removes some of the boilerplate for both accepting a
connection pool for a new handler struct as well as improves the
interaction of fetching a connection from the pool. It extracts this
code that is duplicated in every handler into a new WithDB trait that
can get a reference to a pool and also get a connection from that pool.
The trait is implemented on the new MicropubDB struct which is now a
part of each handler.

While I was at it, I extracted the transaction functionality that the
micropub handler used into the WithDB trait so that eventually other
handlers could use it.