refactor: avoid n+1 tag lookup in atom and archive handlers
This refactors the handler code to look for tags for all post IDs in one
query rather than one at a time. In theory this should be faster though
I didn't see any improvement on a release build locally. It could be
that it wasn't much of a problem or just that my machine is fast enough
to paper over the problem (both of those pages load 3x or more faster
locally than the server my site runs on). Committing both because in
theory it's faster anyway and because I'd like to run it on the web
server to see the result.

A nice follow up would be to extract this logic out so it doesn't live
in both places. A bunch of the post lookup and Post view model creation
could be generalized as it's used in almost every handler. (Though only
these two need to do tag lookups for multiple posts at once).