feat: add bookmark-of support for bookmark style posts
This commit implements the bookmark-of micropub property. See
https://indieweb.org/bookmark for details but the gist is that some
clients support creating a "bookmark" with a title, bookmark-of url,
optional content that may or may not be sourced from the bookmark page.

It implements bookmarks in the atom feed (not linking directly to them
but rather including a little bookmark emoji in the title and content
that will link to the bookmark) and provides a bookmark_of field in the
view model so that templates might use it.