feat: support micropub config query
Adds support for GET requests against the /micropub path as long as a
q=config is passed in the query string. This lays the groundwork for
both a media endpoint (the response contains the media-endpoint value)
as well as further query functionality (other types of
/micropub?q=<whaterver> queries are possible).

See the Micropub spec, sections 3.6.1 and 3.7 for more details.

In order to support returning the media endpoint, a new configuration
environment variable is expected: MICROPUB_RS_MEDIA_ENDPOINT. The value
is expected to contain the full path to return as part of the q=config
response as the media-endpoint property.