feat: support media upload and fetching
This commit is a first pass implementation of uploading media using a
media endpoint in the micropub protocol. It also adds fetching of the
media with a GET endpoint that takes precedent over the post fetch
handler. It will return the data as well as the content-type sent by the
original upload (or else treat the content type as bytes).

This began months ago and so I don't remember all the details now that I
finally have this working.

It introduces the use of an external service for dealing with storing
the media: rustyblobjectstore. This allows this server to decouple
storing metadata about the media (content type, content hash,
updated/created at time) from the storage of the actual content -
without tying us to any particular storage backend directly.

This isn't super relevant to the implementation here but, for now,
rustyblobjectstore just stores objects sent to it locally in a sqlite
database. This is a bit of an experiment but it wouldn't be too hard to
modify that service to upload images to S3 or something instead. (One
downside of the current model where media is not a separate subdomain is
that we'd still have to proxy the fetch of the images on S3 or whatever
other external storage service we used - not a huge deal and something
we could address when we get there but still worth recording. It would
mean we don't save on bandwidth or request handling - just storage on
the local instance.). This sort of transition would be entirely
transparent to the micropub server though and no modifications would be
needed within this server!

The fetch side is fairly straightforward

Some things that might be improved down the road:

- We use Streams in a few places (multipart upload to micropub server as
  well as fetching the media back from the object store). Currently we
  buffer the media in memory before sending the body where it's going. It
  would be great to be able to stream the contents directly without
  buffering. As I don't (currently) anticipate uploading many large
  video or other files, this shouldn't be a big concern. But even with
  photos could be problematic if there are many concurrent requests for
  media. (Another drawback of media being proxied through this server...)
- See below about 'photo' properties in the form body:

Additionally, I've tested with both the Editor and Note tools on
https://quill.p3k.io: The upload functionality currently works with the
Editor (to create a Post/Article) but does not work with the Note
creation tool. This is because that tool sends the photo as a photo
entry in the micropub form body rather than as an html img tag in the
body of the post.

For now, I thought that getting this committed as-is would be good and
could add support for the photo property of the form body as a separate
improvement.