Implement micropub photo property
This implements the ability to add photos to entries even if the content
of the entry is plain-text. This is via adding photo URLs to the entry
via the 'photo' property on the entry.

This builds on the previous addition of media uploads: One micropub
client (quill.p3k.io) happens to use the photo property even for photo
uploads in the case that the post is a 'note'.

This commit adds a photos table for tracking photos added to posts. It
also handles adding items to that table on post creation as well as
fetching them on reads of posts.

I've skipped reading photo entries for the archive page because they are
not displayed currently anyway.

One future improvement might be to fetch and store the photo within the
micropub/rustyblobjectstore system if the URL does not match a media URL
of the current server. The spec suggests this is a thing servers may do
while it is not required to be conforming.