feat: Strip EXIF tags when media is identifiable
On uploads, we attempt to identify the format of the upload. If we can
identify it as an image we attempt to strip tags using ImageMagick
(includes Exif but other tags as well) and will fail the request if we
cannot.

As a way to try to strip tags in more cases, we still attempt to strip
tags even if we do not identify the upload format as an image. In this
case, we don't fail the upload if stripping fails.

Also adds nix configuration for building the project as well as building
a docker image. With the previous Dockerfile, I ran into issues making
the correct ImageMagick symbols available to the server at runtime and
it was not stripping tags. The nix-built Docker image does not have that
problem.