Fix for url slugs with slashes
e.g. 2022/12/27/some-post-name

Previously this wasn't matching because the route (whether :slug_url or
a wildcard like *slug_url) wouldn't quite work.

The former didn't work because it only matches upto the first slash
(2022, in my above example). The latter didn't _seem_ to work because of
a panic with router overlap with all the other routes since I was doing
the wildcard at the root level of the router.

Fallback seems to work ok!