Make micropub-rs configurable via toml file
I've long wanted to move all the "configuration" that was hardcoded in
constants and loaded via environment variables into a single central
toml file for easier documentation and modification.

This commit does just that: it does away with all runtime environment
variables that are currently required as well as any constant values
that were hardcoded values that should be configurable.

This brings micropub-rs one step closer to being agnostic as to what
domain it serves (rather than only being tied to my personal site).

Not currently including documentation on the config structure but I hope
to do that soon.