Move tasks over to rayon thread pool + adjust db connection pool
The rayon threadpool alone was not enough to solve perf problems with
offloading these tasks. I had thought it was related to queuing or too
much parallelism and it (kinda) was but really it was exhaustion of
the db connection pool (default 10) with a high timeout (30s).

These two adjustments combined now at least prevent the server from
going unresponsive when I hit it with 5000 rps. The server _before_
any offloading (just running all this on the Tokio runtime threads)
is still lower latency. It's probable that this "blocking/CPU work"
is just still so little the thread coordination overhead is higher
than the work at hand.